Add a successful collective scheduling test
