+++ apps/web/test/utils/bookingScenario/bookingScenario.ts
@@ -122,7 +122,7 @@ type InputBooking = {
   }[];
 };
 
-const Timezones = {
+export const Timezones = {
   "+5:30": "Asia/Kolkata",
   "+6:00": "Asia/Dhaka",
 };
@@ -866,42 +866,50 @@ export function mockVideoApp({
         lib: {
           // eslint-disable-next-line @typescript-eslint/ban-ts-comment
           //@ts-ignore
-          VideoApiAdapter: () => ({
-            // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            createMeeting: (...rest: any[]) => {
-              if (creationCrash) {
-                throw new Error("MockVideoApiAdapter.createMeeting fake error");
-              }
-              createMeetingCalls.push(rest);
+          VideoApiAdapter: (credential) => {
+            return {
+              // eslint-disable-next-line @typescript-eslint/no-explicit-any
+              createMeeting: (...rest: any[]) => {
+                if (creationCrash) {
+                  throw new Error("MockVideoApiAdapter.createMeeting fake error");
+                }
+                createMeetingCalls.push({
+                  credential,
+                  args: rest,
+                });
 
-              return Promise.resolve({
-                type: appStoreMetadata[metadataLookupKey as keyof typeof appStoreMetadata].type,
-                ...videoMeetingData,
-              });
-            },
-            // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            updateMeeting: async (...rest: any[]) => {
-              if (updationCrash) {
-                throw new Error("MockVideoApiAdapter.updateMeeting fake error");
-              }
-              const [bookingRef, calEvent] = rest;
-              updateMeetingCalls.push(rest);
-              if (!bookingRef.type) {
-                throw new Error("bookingRef.type is not defined");
-              }
-              if (!calEvent.organizer) {
-                throw new Error("calEvent.organizer is not defined");
-              }
-              log.silly(
-                "mockSuccessfulVideoMeetingCreation.updateMeeting",
-                JSON.stringify({ bookingRef, calEvent })
-              );
-              return Promise.resolve({
-                type: appStoreMetadata[metadataLookupKey as keyof typeof appStoreMetadata].type,
-                ...videoMeetingData,
-              });
-            },
-          }),
+                return Promise.resolve({
+                  type: appStoreMetadata[metadataLookupKey as keyof typeof appStoreMetadata].type,
+                  ...videoMeetingData,
+                });
+              },
+              // eslint-disable-next-line @typescript-eslint/no-explicit-any
+              updateMeeting: async (...rest: any[]) => {
+                if (updationCrash) {
+                  throw new Error("MockVideoApiAdapter.updateMeeting fake error");
+                }
+                const [bookingRef, calEvent] = rest;
+                updateMeetingCalls.push({
+                  credential,
+                  args: rest,
+                });
+                if (!bookingRef.type) {
+                  throw new Error("bookingRef.type is not defined");
+                }
+                if (!calEvent.organizer) {
+                  throw new Error("calEvent.organizer is not defined");
+                }
+                log.silly(
+                  "mockSuccessfulVideoMeetingCreation.updateMeeting",
+                  JSON.stringify({ bookingRef, calEvent })
+                );
+                return Promise.resolve({
+                  type: appStoreMetadata[metadataLookupKey as keyof typeof appStoreMetadata].type,
+                  ...videoMeetingData,
+                });
+              },
+            };
+          },
         },
       });
     });
