+++ packages/features/bookings/lib/handleNewBooking/getMockRequestDataForBooking.ts
@@ -1,0 +1,21 @@
+import { getBasicMockRequestDataForBooking } from "./getBasicMockRequestDataForBooking";
+
+export function getMockRequestDataForBooking({
+  data,
+}: {
+  data: Partial<ReturnType<typeof getBasicMockRequestDataForBooking>> & {
+    eventTypeId: number;
+    rescheduleUid?: string;
+    bookingUid?: string;
+    responses: {
+      email: string;
+      name: string;
+      location: { optionValue: ""; value: string };
+    };
+  };
+}) {
+  return {
+    ...getBasicMockRequestDataForBooking(),
+    ...data,
+  };
+}
